extract MainContainer component from app
